requirements.txt
pandas
sqlalchemy
mysql-connector-python
python-dotenv
requests
